requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
pyjwt==2.8.0
python-dotenv==1.0.0
requests==2.31.0